`all_open_transactions` should not include invalidated transactions
invalidated transactions aren't really open. This can happen when
a [transaction is retried][1], so the transaction is invalidated and
when executing the block again, we would get an exception.

As there is no open transaction, the `after_all_transactions_commit`
blocks should just execute as if there were no open transactions.

[1]: https://github.com/rails/rails/blob/d3e3eefeef20bfbec69a8358a7c39d3a7f20574c/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb#L1032