Change the internal transaction state to be closed when the transaction is finalized
Before we were using the null pattern to calculate if the transaction
is open or not, but there are cases like when the transaction is invalidated
that we keep the transaction object around, and just change the state.

The previous behavior made us miss cases where the transaction aren't
not really open in our code. See 2f571a5258278f66cb8ef05e667fd7ff1d3d0fe5.

The public facing transaction object already had the behavior we are
implementing here, so we are just aligning the internal transaction
object to have the same behavior.