Disconnect existing SQLite connection before dropping database
This prevents leaving the connection in a bad state after removing the
DB files, and then using the connection for something else that might
not work. This happens if you have only `development` configured and you
run the tasks `db:reset` or `db:drop` followed by other tasks like
`db:create db:schema_load`, etc. The same connection is reused for tasks
after `db:drop`, and that won't work. It doesn't happen when both
`development` and `test` environments are configured, as DB tasks for
those are run together, and switching between dev and test ensures we
disconnect from the connection between tasks.

Fixes #53832