Mysql2Adapter: gracefully handle `Unknown prepared statement handler` errors
The source of this error is really unclear, I suspect `mysql2` or `libmysqlclient`
sometimes silently reconnect, which cause all handlers to be invalidated, but
I have no proof of that.

However, if we get this precise error code, we know it's safe to retry
as there was no side effect.