Extract --skip docs into section
Previously, the CLI docs describing `--skip` options were a
mostly-comprehensive list of options available that basically duplicates
the output of `rails new --help`. A trickle of PRs have been opened to
add missing options to the list, but it remains incomplete.

This isn't ideal because the list is toilsome to keep in sync with the
actual list of options (and keeping them in sync isn't
enforced/automatic). Additionally, the list isn't providing much
independent value because the end of the section directs users to
`rails new --help` for a full list of options anyways. 

This commit aims to address these shortcomings by turning the content
about `--skip` options into a dedicated section with an example. The
example demonstrates using a `--skip` and how it can cause other things
to be skipped as well. Additionally, the skip section is moved below
"Preconfigure a Different Database" because that feels like a better
introduction to CLI options than `--skip`.