Allow to tag serialized attributes as comparable
Fix: https://github.com/rails/rails/pull/28416

A not rare issue when working with serialized attributes is that
the serialized representation of an object can change over time.
Either because you are migrating from one serializer to the other
(e.g. YAML to JSON or to msgpack), or because the serializer used
subtly changed it's output.

One example is `libyaml` that used to have some extra trailing
whitespaces, and recently fixed that: https://github.com/yaml/libyaml/pull/186

When this sorts of thing happen, you end up with lots of records that
report being changed even though they aren't, which in the best case
leads to a lot more writes to the database and in the worst case lead to
nasty bugs.

So ideally rather than to compare serialized representations, we'd
compare the deserialized objects, however we can't assume the deserialized
type has a proper working `==` method that deeply compare objects.

So the best thing we can do is to only do this if the type is declared
as being comparable.