Return a 500 on MissingController errors
Fix: https://github.com/rails/rails/issues/14567

A controller not existing isn't a routing error that should result
in a 404, but a programming error that should result in a 500 and
be reported.

Until recently, this was hard to untangle because of the support
for dynamic `:controller` segment in routes, but since this is
deprecated and will be removed in Rails 8.1, we can now easily
not consider missing controllers as routing errors.