Fix guides url does not work when clicking the link [ci skip]
If you click the current link in the guides, it will open this page `https://guides.rubyonrails.org%2A%2A`
because it is taking the `*` as part of the url, which is not correct.

Initially, I only added an space between the link and the `*` to fix the issue, but the markdown was not rendering correctly.