Warn to set config.as.variant_processor = nil
Fixes #54152

In #45100 we decided that when the variant_processor is :vips or
:mini_magick, we should try to eagerly load the library and throw a
warning if the library was not available.

However, because `:vips` is the default, this will always show the
warning unless the developer adds `image_processing` gem to their
dependencies.

If they do not want to do this, they can opt-out by setting
`config.active_storage.variant_processor = nil`.