Add `railspect requires` for auditing `require`s
This first commit sets up the command, a visitor to gather data, and a
simple lint to demonstrate its use.

This first lint ensures that "active_support/rails" is required in all
frameworks (one was missing this require but this was addressed in a
[previous commit][1]), and then checks that none of the requires in
"active_support/rails" are called again (except in Active Support since
it does not require "active_support/rails").

[1]: 2e920b0097dab768738ea20849dec8029f31e14e