Fix ActionDispatch::Request::HTTP_METHOD_LOOKUP being impacted by inflections from gems
If a gem adds inflections before action_dispatch/http/request is loaded into the runtime,
then it's possible that calls to method#underscore cause invalid values to be generated.

For example, if a gem adds "OS" as an acronym, then this will cause "POST" to be converted
to :p_os_t.