Enable `#exists?` queries to be retryable
Previously, `#exists?` queries could never be retryable because `"1 AS
one"` was passed as a String (which would be wrapped as a non-retryable
`SqlLiteral`).

This commit addresses that issue by wrapping `"1 AS one"` in a
`SqlLiteral` with `retryable: true`. While not guaranteeing that all
`#exists?` queries can be retryable, this no longer prevents otherwise
retryable `#exists?` queries from being retryable.

Additionally, a small change was needed to ensure the retryable
`SqlLiteral` wasn't wrapped in a non-retryable `SqlLiteral`. A [previous
commit][1] added a `to_s` in `arel_column` which caused all
`SqlLiteral`s to be non-retryable (the `Arel.arel_node?` predicate would
always be false since `field` was always a `String`). The `to_s` was
moved to the specific case it was needed, which was inside `sum`
calculations.

[1]: aaad4932b26d731d9aa2b14f21207d6ddf83ea73

Co-authored-by: classidied <claire.wang028@gmail.com>