Remove unused routes from auth generator
When you run bin/rails generate authentication, the generator creates a
`resource` route for `sessions` and a `resources` route for passwords.
The generated controllers only implement a subset of the needed resource
routes.

Modify the AuthenticationGenerator so that it only generates the
resource routes that are used by the SessionsController (new, create,
destroy) and PasswordsController (new, create, edit, update).

Fixes #54233