Add a script for using dev containers outside VSCode
I want to use the devcontainer info, but I don't want to use VSCode.
This commit adds a small script that reads the devcontainer JSON file
and runs docker commands based on the information in the JSON file.

Usage is like this:

In terminal 1, run `tool/devcontainer up`

After that thing is running do `tool/devcontainer run-user-commands` to
make sure gems are bundled and database is set up.

Finally run `tool/devcontainer sh` to get a shell inside the container.