Make `raise_on_missing_translation` ignore `i18n.plural.rule`
Make `raise_on_missing_translation` ignore `i18n.plural.rule`
because when the `I18n::Backend::Pluralization` module is included
it will look for a pluralization rule override as a translation lookup
at `i18n.plural.rule` for the current locale:
https://github.com/ruby-i18n/i18n/blob/v1.14.6/lib/i18n/backend/pluralization.rb#L82

If a rule is missing for the locale, then default logic is used.
However, if `raise_on_missing_translation` is enabled
the missing rule causes an uncaught exception inside the `i18n` gem.
This exception will bubble up to the app with an error message
seemingly unrelated to the requested translation.

This fixes a regression when upgrading from Rails 7.0 to 7.1
for applications that meet all of the following criteria:
- `raise_on_missing_translations` is set to true.
- A pluralization translation is used by passing `count` option.
- `i18n.plural.rule` is not set for the current locale.
  This is not set by default in a Rails application
  and it does not need to be set if the default logic
  in the `i18n` gem works as expected for the current locale.
- `Backend::Pluralization` is included into `I18n::Backend::Simple`.
  This does not happen by default in a Rails application.
  The Rails I18n guide shows it as part of a customization example:
  https://guides.rubyonrails.org/v7.1/i18n.html#locale-specific-rules
  and some gems execute this module inclusion automatically.