Allow changing the command used for `db:schema:dump` and `db:schema:load` when using sql format for schema.
This can be overridden by setting
`ActiveRecord::Tasks::DatabaseTasks.structure_dump_command` or
`ActiveRecord::Tasks::DatabaseTasks.structure_load_command`
respectively.

The use-case for this flag is to support overriding the default command
used for structure dumping and loading to point to a different (newer)
version than exists in one's path.

This situation can occur when supporting multiple databases or versions
of databases in the same, or across, projects. The need to juggle the
path manually per-project is the main pain point.

Co-authored-by: Hartley McGuire <skipkayhil@gmail.com>