Consolidate database commands to config
We added database configuration for custom structure dump/load commands,
to be configured per database.

The use-case for this is if you have multiple databases using different
versions, or in the process of a database version upgrade and you need
multiple versions of the command installed. This can make managing them
in your path complicated.

We also added configuration for custom dbconsole command, this was
already under the `ActiveRecord.database_cli` option.
However, that option doesn't support multiple versions of the same
database like we wanted for structure dump/load. Having the default
hard-coded on a top-level constant like AR with the sole-purpose of
servicing the `dbconsole` command felt misplaced. With the hash being
keyed by the database type, and not the adapter is especially confusing
when other configurations are adapter based. Finally, because it's a
simple hash, deprecating it or changing it are extremely difficult so
it's better to do that while the option is still new (8.0) and only
used by dbconsole.

Following the structure dump/load commands, we also added flags to
database config and deprecated the old setters.

Co-authored-by: Hartley McGuire <skipkayhil@gmail.com>