PoolConfig double-checks its connection class reference for reloading
Keeping a reference to the class caused subtle issues when combined
with reloading in development. Now, we take care to try to re-acquire
the class reference by constantizing its name.

That lookup may fail in scenarios where classes are given names but
are not assigned to a constant. In that case, the class can't/won't be
reloaded, so we're not worried about it.

Fixes #54343.