Make the schema cache forward compatible with the changes in Column
Since #54333 The Column class added a new attribute `cast_type` that
might not be in the schema cache dump made by a Rails 8.0 application.

In that case, we should fallback to the previous behavior and fetch the
type from the connection.

Of course this is not ideal since it will still require a connection to
get the cast type, but for schema cache dumps made by Rails 8.1 we will
not need to connect to the database anymore.

This code can be removed after we release Rails 8.1.