Do not set the `Clear-Site-Data` header for Chrome users
Since d51e570, the session controller generated by the authentication
generator sets the `Clear-Site-Data` header on logout so that the
browser clears the storage and cache data associated to the site.
Unfortunately, this takes [20+ seconds][1] on Chrome. Since clearing
this data isn't critical, let's not set the header for Chrome users.

I considered opting out all Chromium-based browsers, but from my
testing, Edge on macOS is not affected (despite being Chromium-based).
So I assumed that only Chrome was affected.

[1]: https://issues.chromium.org/issues/41343050