Add enum values option to PostgreSQL enum columns
This is done so that you can specify the values of an enum
inline when defining a column. To prevent confusion between
enum type naming, an error would raise if an existing enum type is
being created with different values. If an enum_type is not
specified, the enum type will default to the column name.
Schemas will include the values of the enum type in eachcolumn definition.

    def up
      create_table :cats do |t|
        t.enum :current_mood, enum_type: mood, values: [happy, sad], default: happy, null: false
      end
    end