Use refactored Timezone interface in public API
The previous code relied on rescuing an exception to produce an empty timezone lookup. That code did not distinguish between types of exceptions, so it could hide an implementation error for a valid timezone. In addition, exception based flow control is (usually) slower (however I've not benchmarked the difference here, just noting it as an extra, expected, benefit).