Support externally divergent APIs
Prior code was failing due to the refactor:

```
$ bin/test test/time_zone_test.rb
$ bin/test test/core_ext/time_with_zone_test.rb
```

`Time.find_zone(Object.new)` is expected to return `nil` while `ActiveSupport::TimeZone[false]` is expected to raise an ArgumentError. This is somewhat surprising as the semantics of both would seem to be the same at a glance. However, the behavior is surely relied on by applications at this point and the goal of this PR isn't to change existing behavior, but to add additional information on error.