Restore Active Storage config to disable variants and analyzers
Fixes a couple of mistakes made in e978ef0.

* Allow setting `config.active_storage.analyzers = []`
* Allow disabling `config.active_storage.variant_processor` to remove
  warning
* Ensure original default `ActiveStorage.analyzers` array is not broken
  (all available analyzers)
* Ensure original default `ActiveStorage.variant_processor` is not
  broken (`:mini_magick` until `load_defaults(7.0)`)

This change restores the configuration of analyzers back to 8.0, instead
of trying to gracefully handle loading those constants if the associated
gem is missing.

Due to that config being set at require time, this happens prior to
initialization and we don't have access to a logger or deprecator to
warn the user.

Since `image_processing` gem transitively depends on `ruby-vips` and
`mini_magick`, you only need to install that to remove the warning.

However, if you're not using variant transformers, but still using the
default analyzers you can remove the warning by including those gems or
setting the analyzers config to an empty array.

We determine the variant transformer based on
`config.active_storage.variant_processor` value, or nothing if you
passed an invalid value (like `mmmmmini_magick`).

Passing an invalid value, should have the same existing behavior and not
impact a working application but in the future we'd like to deprecate it.

Co-authored-by: Alexandre Ruban <alexandre@hey.com>