Explicitly call I18n.exception_handler on missing translation
Fix: https://github.com/rails/rails/issues/54419

Since `TranslationHelper#translate` always sets a `default`,
missing translation is never considered like an error by `I18n`
and the error handler is never called.

This makes I18n's exception handler pretty much unusable.

The solution is to explictly call it ourselves, but the downside
is we need to allocate that exception, which isn't free, but
supposedly it's not expected to happen often.