Make enum column definitions db-agnostic
This is done by introducing a values option to specify the values of an enum for MySQL adapters.
Using the values option in PostgreSQL would allow implicit creation of enum types without
create_enum. If no enum_type was specified, the name of the enum will default to the column
name. Enums in SQLite are represented as strings. Loading a schema with enum columns are
now supported with any adapter.

    def up
      create_table :cats do |t|
        t.enum :current_mood, enum_type: mood, values: [happy, sad], default: happy, null: false
      end
    end