Enable statement-cached queries to be retryable
The StatementCache uses a different code path for generating SQL and
executing queries, so it doesn't currently benefit from the ability to
mark queries as retryable. This mainly affects association SELECTs,
because the other user of StatementCache is `find_by` and it explicitly
passes `allow_retry: true` when executing a cached query.

This commit enables queries that would be retryable without using the
StatementCache to also be retryable when using it. Just like in
`to_sql_and_binds`, `cacheable_query` now sets `retryable: true` on the
`collector` that constructs the query and additionally copies the
`collector`'s final `retryable` value to the query object. Then, instead
of accepting an `allow_retry` parameter when executing the cached query,
`execute` now uses the query object's `retryable` attribute.

The test for this case uses Author/Books instead of Post/Comments
because Comment has a `default_scope`, so `post.comments` does not use
the statement cache.