Make internal query() retryable for postgresql
In a [previous commit][1], the abstract adapter's `query` method was
changed to automatically retry (since the method is used for SCHEMA type
queries, which are idempotent SELECTs). However, the postgresql adapter
has its own `query` method and it was not changed.

This commit makes the same change to the postgresql adapter's `query` so
that it can also benefit from automatically retryable SCHEMA queries.

[1]: b4c19d038e558d12c3e13317759ca13587c41a16