Fix migrating multiple DBs with pending migration action
This was supposed to work with multiple DBs, but the task code and the
migration code have fallen out of sync and the pending migration action
only works against primary. With this patch, we use
`ActiveRecord::Tasks::DatabaseTasks.migrate_all` for both, and introduce
`ActiveRecord::Tasks::DatabaseTasks.dump_all` to use in the task and
action.