Add `allow_retry` to `sql.active_record`
This enables identifying queries which are and are not automatically
retryable on connection errors.

A few end-to-end tests were left using `remote_disconnect` to ensure
that reconnection is covered, but tests that are just trying to assert
that certain queries will cause a reconnect can now make assertions
using `allow_retry` instead. (In the case of Mysql2/Trilogy, this speeds
up tests by 1 second per `remote_disconnect` call removed)