Enable query-cacheable queries to be retryable
The query cache implementation calls `to_sql_and_binds` inside its
`select_all` override so that it can use the generated sql and binds as
a cache key. However, the current implementation does not pass the
`allow_retry` parameter into `to_sql_and_binds`. This is an issue when
the `arel` parameter is a String, because query cache's `select_all`
will override the `allow_retry` parameter with the value returned from
`to_sql_and_binds` (and because it isn't passed through, it will always
be the default, false).

This commit fixes the issue by passing the `allow_retry` parameter
through to `to_sql_and_binds`. This will ensure that when `select_all`
is called with a String query, `allow_retry: true`, and the query cache
enabled that the query will maintain its retryable state.

This combination of prerequisites can occur when a query is executed
from the Statement Cache since the Statement Cache does its own query
compilation.