Fix RDoc for ActionView::TestCase::Behavior#rendered
RDoc was previously treating `rendered` as an attribute due to its definition with `attr_accessor`, preventing it from being documented  as a method.

To resolve this, the documentation was moved above the `attr_accessor` declaration, and `attr_accessor :rendered` was defined separately.  This ensures that `rendered` is no longer listed under attributes and is properly documented as a method.