Ensure allow_retry passes when mysql2 prepared
In Rails 8, adapters were [refactored][1] so have a more uniform
interface, and this included unifying the checking out of a
`raw_connection` for the prepared and unprepared statement code paths.
However, the code paths are distinct on Rails 7.2 and `allow_retry` is
not passed to `raw_connection` in the prepared statement path.

This commit passes through `allow_retry` so that prepared statements can
also be automatically retried.

[1]: fd24e5bfc9540fc00764a59ddf39a993bbd63ba2