Remove test/dummy from ActiveStorage
The goal here is to remove the committed Rails app that lives in
test/dummy used for initializing Rails and thus ActiveStorage.

Instead, replace it with a test setup only required to load the parts it
is testing.

By using the Rails namespace as a mock for the application for two main
reasons:

1. House the routes, but that should be fixed in a
   future PR that makes them modular
2. Some migrations depend on Rails config to determine the primary key
   type, there is no public API to get that info since it's applied
   directly to the migration generator during initialization

The plan would be to refactor any dependencies, both in test and
library, that depend on the Rails namespace to use an appropriate
framework API or introduce one if necessary.

Alternatives considered were _actually_ initializing a Rails app, or
generating one dynamically but that is what Railties does. And we should
be able to test most-if-not-all of the engine without Rails, like we do
with ActionController, ActiveJob, etc.