Micro optimize route sorting in Router find_routes
`Array#sort_by!` is a bit slower than explicit `sort!`,
but more importantly, if we assume we have only matched a single
route, we can save entering the sort method at all.

You'd think `#sort_by!` would bail out early on arrays with
a single item, but that's not the case.