Avoid `#any?` since `via` is usually singular
`Route#matches? currently calls `#any?` to match the Route's array of
request method matchers against the request. Even though the array of
matching request methods is often only one element, every Route ends up
paying the cost of `#any?`.

This commit addresses this cost by specializing the matcher to always be
a single object. For Routes with only one matching request method, the
matcher can be called directly. For Routes with multiple matching
request methods, a new Matcher was introduced to iterate the matchers as
before.