Reorder equality so happy path hits `Nil#==`
For most requests the expectation is that the application will respond
to the request and not set the `x-cascade` header. Since the header will
be `Nil` in that happy path, we should prefer calling `#==` on the
header itself because `String#==` will try to coerce `other` `to_str`,
whereas `Nil#==` will not.

(`String#==` showed up as ~2.5% in a profile, and this eliminates it)