doc: Explicitly state that modifying sanitizer allowlists is unsafe
The allowlists in rails-html-sanitizer are curated to balance security
and usability. Because Rails is a sharp knife, it's possible to add to
these allowlists to suit specific edge cases.

However, doing so may introduce security vulnerabilities into an app,
and this updates the documentation to be clear that Rails cannot
protect them if they insist on allowing unsafe tags or attributes.

Hopefully this change also encourages folks to open an issue if the
allowlists are missing some obvious tags or attributes, so that rather
than individual developers modifying the allowlists, we can improve
the functionality for everyone.

[skip ci]