Fix LocalStore#read_multi_entries to distinguish recorded misses
When reading a missing key, the local store set that key to `nil`,
so we know not to query the real cache later on.

However when using `fetch` or `fetch_multi`, we should treat it as
a miss and invoke the block.

So at that point we might as well behave as we didn't know it was
a miss.