Only create one config per AV::Base
Currently, AV::Base's config is always initialized to a blank
InheritableOptions, and then later config is overridden with the given
controller's config if it responds to config (which it will most of the
time since AbstractController::Base has a config).

This commit moves the blank config instantiation into the
ControllerHelper so that only a single InheritableOptions is ever
instantiated for an AV::Base.