Reset execution context in jobs and controllers
Currently `ActiveSupport::ExecutionContext#[]` is used to set the `:job`
and `:controller` attributes, which does not reset the values once the
job or controller have completed execution. This can result in that
context leaking, resulting in query logs that are misattributed.

This leaky behavior can be seen in production (e.g. performing a job via
`perform_now`) but is much more noticeable in the test environment where
it's not uncommon to run some combination of jobs/controller actions in
a single test.

This updates both uses of `ActiveSupport::ExecutionContext#[]` to avoid
leaking the context and generating inaccurate query logs.