Use dedicated GitHub Action jobs for tests and system tests
The most efficient way to speedup a CI is to run the various steps
in parallel. The difficulty is of course hpw to split the workload
in a balanced way.

But regular tests and systems tests are a very natural split strategy,
first because system tests require some extra setup, and are generally
slower, but also because they tend to be more flaky.

So in case of a flake you can retry alone without having to retry
all your other tests. this improve the UX a lot.