Use MySQL instead of MariaDB in devcontainer
MariaDB was likely chosen because Debian "bookworm" sets it as the default
MySQL-compatible database. However, recent changes in MariaDB introduced
incompatibilities with MySQL—for example, #53727. This makes MariaDB less
suitable as the default for developing the mysql2 and trilogy adapters.

Now that the database runs as a separate container, we are no longer bound
by the base OS default. Using the official MySQL container offers a development environment
that more closely aligns with the behavior of MySQL itself.

References:
https://lists.debian.org/debian-devel-announce/2016/09/msg00000.html
https://github.com/rails/rails/pull/48891