Bump bundler inside the devcontainer
This commit updates the Bundler version in Gemfile.lock
and modifies the boot.sh script to install the latest Bundler version
to avoid similar warnings in the future.

- Steps to reproduce
1. Startup the devcontainer
2. Run
```bash
$ bundle exec rake test
```

- Actual result
This results in the following warning:

```
/home/vscode/.rbenv/versions/3.4.2/lib/ruby/gems/3.4.0/gems/bundler-2.5.16/lib/bundler/rubygems_ext.rb:250: warning: method redefined; discarding old encode_with
/home/vscode/.rbenv/versions/3.4.2/lib/ruby/3.4.0/rubygems/dependency.rb:341: warning: previous definition of encode_with was here
```

- Background:

When Ruby is upgraded, the gem command is updated as well.
However, Bundler installs the version specified in Gemfile.lock under BUNDLED WITH,
which can cause outdated versions and trigger the warning above.

The fix in rubygems/rubygems#7867 has been available since RubyGems 3.6.0 and Bundler 2.6.0.
Since Bundler 2.5.16 lacks this commit, upgrading is necessary to prevent these warnings.